+++ lib/src/internal/editor.dart
@@ -14,14 +14,6 @@ class Editor {
   final DarwinEditor _darwin = const DarwinEditor();
   final AndroidEditor _android = const AndroidEditor();
 
-  @Deprecated('Use `Editor.darwin` instead. This will be removed in 3.0.0')
-  IosEditor get iOS {
-    if (Platform.isIOS || Platform.isMacOS) {
-      return _ios;
-    }
-    throw const OSError('iOS Editor should only be use on iOS.');
-  }
-
   /// Support iOS and macOS.
   DarwinEditor get darwin {
     if (Platform.isIOS || Platform.isMacOS) {